Disable java download for now, add more libraries

+++ ubuntu/systemwide_setup.sh
@@ -2,9 +2,9 @@
 
 set -e -u -o pipefail -x
 
-# -------------------------------------------------------------------------------- 
+# --------------------------------------------------------------------------------
 # Functions
-# -------------------------------------------------------------------------------- 
+# --------------------------------------------------------------------------------
 
 # Check if the given tarball is OK.
 check_tarball() {
@@ -27,6 +27,7 @@ packages=(
 curl
 git
 libxml2-dev
+libxslt1-dev
 libyaml-dev
 openssh-server
 vim
@@ -44,35 +45,38 @@ if [ -f /etc/NetworkManager/NetworkManager.conf ]; then
   sudo restart network-manager
 fi
 
-# Oracle JDK
-jdk_tmp_dir="/tmp/oracle_jdk"
-mkdir -p "${jdk_tmp_dir}"
-cd "${jdk_tmp_dir}"
+if [ ]; then
 
-jdk_tarball=jdk-7u51-linux-x64.tar.gz
-if ! check_tarball "${jdk_tarball}"; then
-  rm -f "${jdk_tarball}"
-  wget --no-cookies --header "Cookie: gpw_e24=http%3A%2F%2Fwww.oracle.com" \
-    "http://download.oracle.com/otn-pub/java/jdk/7u51-b13/${jdk_tarball}"
-fi
+  # Oracle JDK
+  jdk_tmp_dir="/tmp/oracle_jdk"
+  mkdir -p "${jdk_tmp_dir}"
+  cd "${jdk_tmp_dir}"
 
-# Get the top-level directory name from the tarball.
-new_jdk_dir_name=$( top_level_dir_from_tarball "${jdk_tarball}" )
-
-# Strip the trailing "/".
-new_jdk_dir_name=${new_jdk_dir_name%/}
-
-jdk_top_dir=/usr/lib/jvm
-jdk_target_dir="${jdk_top_dir}/${new_jdk_dir_name}"
-
-if [ ! -d "${jdk_target_dir}" ]; then
-  sudo mkdir -p "${jdk_top_dir}"
-  cd "${jdk_top_dir}"
-  sudo tar xzf "${jdk_tmp_dir}/${jdk_tarball}"
-  if [ -h "default-java" ]; then
-    sudo unlink default-java
+  jdk_tarball=jdk-7u51-linux-x64.tar.gz
+  if ! check_tarball "${jdk_tarball}"; then
+    rm -f "${jdk_tarball}"
+    wget --no-cookies --header "Cookie: gpw_e24=http%3A%2F%2Fwww.oracle.com" \
+      "http://download.oracle.com/otn-pub/java/jdk/7u51-b13/${jdk_tarball}"
   fi
-  sudo ln -s "${new_jdk_dir_name}" default-java
+
+  # Get the top-level directory name from the tarball.
+  new_jdk_dir_name=$( top_level_dir_from_tarball "${jdk_tarball}" )
+
+  # Strip the trailing "/".
+  new_jdk_dir_name=${new_jdk_dir_name%/}
+
+  jdk_top_dir=/usr/lib/jvm
+  jdk_target_dir="${jdk_top_dir}/${new_jdk_dir_name}"
+
+  if [ ! -d "${jdk_target_dir}" ]; then
+    sudo mkdir -p "${jdk_top_dir}"
+    cd "${jdk_top_dir}"
+    sudo tar xzf "${jdk_tmp_dir}/${jdk_tarball}"
+    if [ -h "default-java" ]; then
+      sudo unlink default-java
+    fi
+    sudo ln -s "${new_jdk_dir_name}" default-java
+  fi
 fi
 
 # Apache Maven
@@ -107,4 +111,3 @@ fi
 if ! egrep "^JAVA_HOME=" /etc/environment >/dev/null; then
   sudo bash -c '( echo; echo "JAVA_HOME=/usr/lib/jvm/default-java" ) >>/etc/environment'
 fi
-
